fix small bugs in productService and reviewService

+++ restaurant-service/src/main/java/com/CNTTK18/restaurant_service/service/productService.java
@@ -146,7 +146,9 @@ public class productService {
         products product = productRepo.findById(id).orElseThrow(() -> new ResourceNotFoundException("Product not found"));
         List<reviews> rv = reviewRepository.findByReviewId(id).stream()
                                 .filter(r -> r.getReviewType().equals(reviewType.PRODUCT.toString())).toList();
-        imageFileService.deleteImage(product.getPublicID());
+        if (!product.getPublicID().isEmpty()) {
+            imageFileService.deleteImage(product.getPublicID());
+        }
         reviewRepository.deleteAll(rv);
         productRepo.delete(product);
     }
